requirements.txt
ExifRead==3.0.0
pillow==10.4.0
